services/requirements.txt
requests
mysql-connector-python
boto3